refactor: Move test_objecttag_export_helpers.py to test directory
